+++ hp3parclient/client.py
@@ -26,8 +26,10 @@ also supports running actions on the 3PAR that use SSH.
 This client requires and works with 3PAR InForm 3.1.3 MU1 firmware
 
 """
+import base64
 import re
 import time
+import uuid
 
 try:
     # For Python 3.0 and later
@@ -1632,9 +1634,24 @@ class HP3ParClient(object):
         # error.  If we don't get that error, we nuke the
         # fake host.
 
+        def _hostname():
+            # create a safe, random hostname that won't
+            # create a collision when findHost is called
+            # in parallel, before the temp host is removed.
+            uuid_str = str(uuid.uuid4()).replace("-", "")
+            safe_uuid = uuid.UUID("urn:uuid:%s" % uuid_str)
+            encoded = base64.b64encode(safe_uuid.bytes)
+
+            # 3par doesn't allow +, now /
+            encoded = encoded.replace('+', '.')
+            encoded = encoded.replace('/', '-')
+            # strip off the == as 3PAR doesn't like those.
+            encoded = encoded.replace('=', '')
+            return encoded
+
         cmd = ['createhost']
         # create a random hostname
-        hostname = 'zxy-delete-vxz'
+        hostname = _hostname()
         if iqn:
             cmd.append('-iscsi')